shooting mechanics

+++ Way_of_Blood/Assets/Scripts/Character/Player/PlayerHealth.cs
@@ -6,7 +6,7 @@ namespace WayOfBlood.Character.Player
     public class PlayerHealth : CharacterHealth
     {
         [Header("Shields parameters")]
-        public float    ShieldsCount;   // Количество щитов
+        public int    ShieldsCount;   // Количество щитов
         public int      ShieldCost;     // Стоимость щита в единицах крови
 
         private CharacterBlood  _characterBlood;
